+++ adm/app/adms/Views/confEmails/listConfEmails.php
@@ -7,7 +7,9 @@ if(!defined('C8L6K7E')){
 
 echo "<h2>Listar Usuários</h2>";
 
-echo "<a href='" . URLADM . "add-conf-emails/index'>Cadastrar</a><br><br>";
+if ($this->data['button']['add_conf_emails']) {
+    echo "<a href='" . URLADM . "add-conf-emails/index'>Cadastrar</a><br><br>";
+}
 
 if (isset($_SESSION['msg'])) {
     echo $_SESSION['msg'];
@@ -22,9 +24,19 @@ foreach ($this->data['listConfEmails'] as $confEmail) {
     echo "Titulo: $title <br>";
     echo "Nome: $name <br>";
     echo "E-mail: $email <br>";
-    echo "<a href='" . URLADM . "view-conf-emails/index/$id'>Visualizar</a><br>";
-    echo "<a href='" . URLADM . "edit-conf-emails/index/$id'>Editar</a><br>";
-    echo "<a href='" . URLADM . "delete-conf-emails/index/$id' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br>";
+
+    if ($this->data['button']['view_conf_emails']) {
+        echo "<a href='" . URLADM . "view-conf-emails/index/$id'>Visualizar</a><br>";
+    }
+    
+    if ($this->data['button']['edit_conf_emails']) {
+        echo "<a href='" . URLADM . "edit-conf-emails/index/$id'>Editar</a><br>";
+    }
+    
+    if ($this->data['button']['delete_conf_emails']) {
+        echo "<a href='" . URLADM . "delete-conf-emails/index/$id' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br>";
+    }
+    
     echo "<hr>";
 }
