+++ adm/app/adms/Views/confEmails/viewConfEmail.php
@@ -7,11 +7,22 @@ if(!defined('C8L6K7E')){
 
 echo "<h2>Detalhes da Configuração de E-mail</h2>";
 
-echo "<a href='" . URLADM . "list-conf-emails/index'>Listar</a><br>";
+if ($this->data['button']['list_conf_emails']) {
+    echo "<a href='" . URLADM . "list-conf-emails/index'>Listar</a><br>";
+}
+
 if (!empty($this->data['viewConfEmails'])) {
-    echo "<a href='" . URLADM . "edit-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Editar</a><br>";
-    echo "<a href='" . URLADM . "edit-conf-emails-password/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Editar Senha</a><br>";
-    echo "<a href='" . URLADM . "delete-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br><br>";
+    if ($this->data['button']['edit_conf_emails']) {
+        echo "<a href='" . URLADM . "edit-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Editar</a><br>";
+    }
+    
+    if ($this->data['button']['edit_conf_emails_password']){
+        echo "<a href='" . URLADM . "edit-conf-emails-password/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Editar Senha</a><br>";
+    }
+    
+    if ($this->data['button']['delete_conf_emails']) {
+        echo "<a href='" . URLADM . "delete-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br><br>";
+    }
 }
 
 if (isset($_SESSION['msg'])) {
