+++ adm/app/adms/Views/pages/viewPages.php
@@ -7,12 +7,18 @@ if(!defined('C8L6K7E')){
 
 echo "<h2>Detalhes da Página</h2>";
 
-echo "<a href='" . URLADM . "list-pages/index'>Listar</a><br>";
+if ($this->data['button']['list_pages']) {
+    echo "<a href='" . URLADM . "list-pages/index'>Listar</a><br>";
+}
+
 if (!empty($this->data['viewPages'])) {
-    echo "<a href='" . URLADM . "edit-pages/index/" . $this->data['viewPage'][0]['id'] . "'>Editar</a><br>";
-    echo "<a href='" . URLADM . "edit-pages-password/index/" . $this->data['viewPage'][0]['id'] . "'>Editar Senha</a><br>";
-    echo "<a href='" . URLADM . "edit-pages-image/index/" . $this->data['viewPage'][0]['id'] . "'>Editar Imagem</a><br>";
-    echo "<a href='" . URLADM . "delete-pages/index/" . $this->data['viewPage'][0]['id'] . "' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br><br>";
+    if ($this->data['button']['edit_pages']) {
+        echo "<a href='" . URLADM . "edit-pages/index/" . $this->data['viewPage'][0]['id'] . "'>Editar</a><br>";
+    }
+    
+    if ($this->data['button']['delete_pages']) {
+        echo "<a href='" . URLADM . "delete-pages/index/" . $this->data['viewPage'][0]['id'] . "' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br><br>";
+    }
 }
 
 if (isset($_SESSION['msg'])) {
