+++ adm/app/adms/Views/sitsUser/listSitsUsers.php
@@ -6,7 +6,9 @@
 
     echo "<h2>Listar Situação</h2>";
 
-    echo "<a href='" . URLADM . "add-sits-users/index'>Cadastrar Situação</a><br><br>";
+    if ($this->data['button']['add_sits_users']) {
+        echo "<a href='" . URLADM . "add-sits-users/index'>Cadastrar Situação</a><br><br>";
+    }
 
     if (isset($_SESSION['msg'])) {
         echo $_SESSION['msg'];
@@ -18,9 +20,19 @@
         extract($sitUser);
         echo "ID: $id <br>";
         echo "Nome: <span style='color: $color'>$name</span> <br>";
-        echo "<a href='" . URLADM . "view-sits-users/index/$id'>Visualizar</a><br>";
-        echo "<a href='" . URLADM . "edit-sits-users/index/$id'>Editar</a><br>";
-        echo "<a href='" . URLADM . "delete-sits-users/index/$id' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br>";
+
+        if ($this->data['button']['view_sits_users']) {
+            echo "<a href='" . URLADM . "view-sits-users/index/$id'>Visualizar</a><br>";
+        }
+        
+        if ($this->data['button']['edit_sits_users']) {
+            echo "<a href='" . URLADM . "edit-sits-users/index/$id'>Editar</a><br>";
+        }
+        
+        if ($this->data['button']['delete_sits_users']){
+            echo "<a href='" . URLADM . "delete-sits-users/index/$id' onclick='return confirm(\"Tem certeza que desaja apagar este registro?\")'>Apagar</a><br>";
+        }
+        
         echo "<hr>";
     }
